Fix requiring of Spring test helpers
The idea of putting these test helpers in the Spring gem was to enable
separate watcher projects to use them like this. That changed with
https://github.com/rails/spring/commit/cfc0c5bce2a1f7da4bfc03d3b0146ab797a9798a.

Since we’re running the tests under Bundler, against a git clone of the
Spring repo (see Gemfile), so we can just reach straight into that.

Closes #24